// ui/types/generated/Vault.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Vault = { id: string, parentVaultId?: string, name: string, icon: string | null, description: string | null, privacyTier: string, priorityProfile: string, summaryNodeId: string | null, sortOrder: number, createdAt: string, updatedAt: string, deletedAt: string | null, meta: string, uiMetadata: string, };